# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.APIKey:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      expires_at:
        type: string
      id:
        type: string
      key_prefix:
        description: sk_live_ or sk_test_
        type: string
      last_used_at:
        type: string
      name:
        type: string
      revoked_at:
        type: string
      revoked_by:
        type: string
      scopes:
        items:
          type: string
        type: array
      status:
        $ref: '#/definitions/domain.APIKeyStatus'
      tenant_id:
        type: string
      type:
        $ref: '#/definitions/domain.APIKeyType'
    type: object
  domain.APIKeyStatus:
    enum:
    - active
    - revoked
    - expired
    type: string
    x-enum-varnames:
    - APIKeyStatusActive
    - APIKeyStatusRevoked
    - APIKeyStatusExpired
  domain.APIKeyType:
    enum:
    - test
    - live
    type: string
    x-enum-varnames:
    - APIKeyTypeTest
    - APIKeyTypeLive
...
  handler.CreateRequest:
    properties:
      expires_at:
        type: string
      name:
        type: string
      scopes:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/domain.APIKeyType'
    type: object
  handler.CreateResponse:
    properties:
      full_key:
        description: The full API key (only shown once)
        type: string
      key:
        $ref: '#/definitions/domain.APIKey'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@saaskit.io
    name: SaaSKit Support
    url: https://github.com/medaminerjb/saas-kit/issues
  description: Enterprise-grade Identity and Access Management platform with OIDC,
    multi-tenant support, and API key management.
  license:
    name: MIT
    url: https://github.com/medaminerjb/saas-kit/blob/main/LICENSE
  termsOfService: https://github.com/medaminerjb/saas-kit/blob/main/TERMS.md
  title: SaaSKit API
  version: "1.0"
paths:
...
  /tenants/{tenantID}/api-keys:
    get:
      consumes:
      - application/json
      description: Get all API keys for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of API keys
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to list
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      description: Create a new API key for a tenant
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: API key details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: API key created
          schema:
            $ref: '#/definitions/handler.CreateResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Create API key
      tags:
      - api-keys
  /tenants/{tenantID}/api-keys/{keyID}:
    delete:
      consumes:
      - application/json
      description: Permanently delete an API key
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: API Key ID
        in: path
        name: keyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: API key deleted
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete API key
      tags:
      - api-keys
    get:
      consumes:
      - application/json
      description: Get details of a specific API key
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: API Key ID
        in: path
        name: keyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: API key details
          schema:
            $ref: '#/definitions/domain.APIKey'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: API key not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get API key
      tags:
      - api-keys
  /tenants/{tenantID}/api-keys/{keyID}/revoke:
    post:
      consumes:
      - application/json
      description: Revoke an API key (soft delete)
      parameters:
      - description: Tenant ID
        in: path
        name: tenantID
        required: true
        type: string
      - description: API Key ID
        in: path
        name: keyID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: API key revoked
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to revoke
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Revoke API key
      tags:
      - api-keys
...
securityDefinitions:
  Bearer:
    description: Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"